Fixed bug @ compilation + input text
There was a bug at compilation that didn't let a function execute. I realized it was not connected to it's parent object, so I fixed that.

I also realized that the Language Specifications input didn't have all tokens defined, so I needed to update that to demonstrate a correct Language Specification by default.